Add test case for spurious rebuild in ddo
I noticed that some systems, e.g. fast-io, will cause UIOP, CFFI, and
everything that depends on it, to be constantly recompiled.
Interestingly, this doesn't happen with any of the systems that fast-io
directly depends on: alexandria, trivial-gray-streams, and static-vectors.
And interestingly, the test case, once reduced, is exactly isomorphic to
the test case already in test-defsystem-depends-on-phase-overlap.script:
fast-io => main-system, alexandria => overlapping-dependency,
static-vectors => intermediate-dependency, bazel => defsystem-dependency
And indeed, adding a test for it detects the failure.