Fix bug detecting change in read files.
Files read in by READ-FILE-LINE or READ-FILE-FORM were not being
tracked.  With this change, they are noticed and added to
ADDITIONAL-INPUT-FILES for a component and operation.

Adding this tracking led to introduction of a new dependency in
asdf.asd.

Also, added a comparison between the operation-time and the latest
update time for the input files of a component and operation.
Previously, even if the input files were newer than the last operation
time, if there were no output files, ASDF would incorrectly think the
operation was up-to-date.