Make the *SHARED-READTABLE* be copy of the standard readtable.
Alternative approach to Fare's syntax control. Fare made this be the
initial readtable.  If this isn't too bad, I would prefer to break this
now, rather than end up needing to be backward compatible with broken
code forever. Especially since the *INITIAL-READTABLE* is not
well-defined (potentially it varies depending on when ASDF is loaded).