Add with-compilation-unit*
A macro that expands to a standard `with-compilation-unit`, wrapped in a
condition handler. The condition handler takes all warnings signaled upon exit
of the wrapped `w-c-u`, collects them in a `deferred-warnings-error` condition
and signals it with `error`.

This behavior can be controlled using the `:asdf-deferred-warnings-behavior`
option to `with-compilation-unit*`, which defaults to the value of
`*with-compilation-unit-deferred-warnings-behavior*`, which itself defaults to
`:error`.