Make COMPONENT-VERSION backwards compatible
COMPONENT-VERSION has explicitly said in its documentation (for several years
now) that the returned version is NIL *or* a string of dot-separated natural
numbers.

In order to maintain backward compatbility, keep the same semantics by only
returning the version if it fits those criteria. To increase the usefulness of
this function, state that if the version is a pre-release (in which case it
almost certainly will *not* meet the criteria), it returns the result of
VERSION-PRE-RELEASE-FOR.

Add function COMPONENT-VERSION* which returns a VERSION-OBJECT directly.

This may be modified if, after checking QL, we determine that no one is
counting on COMPONENT-VERSION having dot-separated natural numbers.