ECL improvements for ADDITIONAL-ENVIRONMENT in LAUNCH-PROGRAM
Warn in situations where LAUNCH-PROGRAM is likely to fail and give reason (old
ECL, ADDITIONAL-ENVIRONMENT provided, not running through a shell, and program
has no directory component). Do not ERROR because dev versions of ECL have this
issue fixed, but still report an old version number.

Improve tests on ECL by forcing a shell on older ECL versions, so that the SHLL
can do the PATH lookup for us.