Summary: Fix ENSURE-DIRECTORY-PATHNAME.
Previously, it set the contents of the `directory` slot of the
pathname using `file-namestring`, but this can introduce escaping into
the pathname object.
On the other hand, we cannot use `pathname-name` alone, because if
there is a `type` in the input pathspec, that must be appended to the
`name` (with an embedded ".").