Add a non-strict option to the ERRORS test.
Previously, `ERRORS` would check that an s-expression raised EXACTLY the
condition type supplied. Now, if one adds `:strict nil`, it's enough
to raise a condition of a subtype of the condition type supplied.

Given how loose the spec is about the set of available condition
classes, it's much easier to check in a non-strict way across
implementations.