Fix SBCL fasl path for sb-thread builds.
Change the logic so that threaded SBCL builds, which is by far the
more common case, does not append anything to the fasl path.

This lowers the probability of an issue where an earlier ASDF is used
to build fasls, then upgraded, and then the new version looks for
fasls in a different directory. This actually happened when building
an executable image for development, which then reinitialized output
translations when it was started, and thus failed to find already
compiled fasls.

Also make sure not to append a dangling dash at the end by not using
an empty string in the false branch.